Reimplement filter system with adabraka-ui sliders
- Replace custom slider implementation with adabraka-ui Slider components
- Implement polling-based change detection architecture
- Fix GPUI image caching with unique timestamped filenames
- Add proper focus management when dismissing filter panel
- Create comprehensive filter system design documentation
- Move design docs to docs/ directory
- Add adabraka-ui dependency to Cargo.toml

This replaces ~300 lines of custom slider code with adabraka-ui components
and fixes issues with focus management and image caching.

+++ src/state/image_state.rs
@@ -52,13 +52,13 @@ impl Default for ImageState {
 }
 
 /// Filter settings for image processing
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct FilterSettings {
-    /// Brightness adjustment (-100 to +100)
-    pub brightness: i32,
+    /// Brightness adjustment (-100.0 to +100.0)
+    pub brightness: f32,
     
-    /// Contrast adjustment (-100 to +100)
-    pub contrast: i32,
+    /// Contrast adjustment (-100.0 to +100.0)
+    pub contrast: f32,
     
     /// Gamma correction (0.1 to 10.0)
     pub gamma: f32,
@@ -67,8 +67,8 @@ pub struct FilterSettings {
 impl Default for FilterSettings {
     fn default() -> Self {
         Self {
-            brightness: 0,
-            contrast: 0,
+            brightness: 0.0,
+            contrast: 0.0,
             gamma: 1.0,
         }
     }
